🏗 Realigns Y2020 Projects for Stronger Typing of Input File
This adds a bunch of stronger-typed mechanisms for working with the input file, with anticipation that these techniques will be equally helpful in upcoming challenges.

The main gist was to have a class with a parameter-less constructor that could provide a working handle to the input file (in the event that we might want to use Xunit and it's "Class Fixture" feature). This in turn led to writing some shortcut extension methods that make working with Task<…> a little less wordy.